Changes n stuff.  Actually added visibility stuff
There's a few interesting things in here, but I'm mega tired.
This was coding this morning at 6.45, with a tailrecusive implementation
of Bresenham's Line Algorithm to calculate visiblity.
We also used memoization to prevent recalculating the visibility a lot.
Finally prevent the player from shooting invisible stuff, and make
visibility more obvious.

Going to sleep now.